Tryed to order the database to make it possible to
calculate specific datasets in the end. This is not necessary
since the getcond.py routine will only be used for calculating
the conductivity in old results with no automatic conductivity
calculation.

Moreover the merging of all possible data combinations with
itertools is to much and very time consuming. This is a dead end.