Introduced long range coupling systems
In addition it is now possible to specify the system type to use
as an input from the database. Hereby it for example possible to
use EuGdO runs as an input for EuGdO-LR runs.

Moreover there is a new help message showing all system types in
the run.py routine.